+++ amplitude.js
@@ -4923,7 +4923,7 @@ module.exports = uuid;
 
 }, {}],
 10: [function(require, module, exports) {
-module.exports = '3.1.0';
+module.exports = '3.2.0';
 
 }, {}],
 11: [function(require, module, exports) {